requirements.txt
amqp==5.0.2
apipkg==1.5
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
billiard==3.6.3.0
black==20.8b1
boto3==1.16.20
botocore==1.19.20
celery==5.0.2
certifi==2020.11.8
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
coverage==5.3
distlib==0.3.1
Django==3.1.3
django-crispy-forms==1.9.2
django-storages==1.10.1
django-tinymce==3.1.0
docutils==0.15.2
execnet==1.7.1
filelock==3.0.12
future==0.18.2
greenlet==0.4.17
gunicorn==20.0.4
identify==1.5.9
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
kombu==4.6.11
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.4
pathspec==0.8.1
Pillow==8.0.1
playwright==0.162.0
pluggy==0.13.1
pre-commit==2.8.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.9.0
pyee==8.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-base-url==1.4.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-playwright==0.0.8
pytest-xdist==2.1.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==1.3.0
virtualenv==20.1.0
wcwidth==0.2.5
